feature/ A lot of changes, mostly everything, but found out that the "relationships" between Tarjetas and Preferencias are not enough, it lacks more relationships with Salario and Edad. Go for rule system

+++ src/main/java/com/rulesAPI/servicios/GenericService.java
@@ -7,9 +7,9 @@ import java.util.Optional;
  */
 public interface GenericService<E> {
     
-    public Optional<E> buscarPorId(Integer id);
+    public Optional<E> buscarPorId(Long id);
     public E guardar(E entidad);
     public Iterable<E> guardarTodos(Iterable<E> entidad);
     public Iterable<E> buscarTodos();
-    public void eliminarPorId(Integer id);
+    public void eliminarPorId(Long id);
 }
